+++ src/methods/$cycle-times/$cycle-times.d.ts
@@ -1,0 +1,7 @@
+import { $SourceIterable, $ResultIterable } from '../../types/$iterable';
+
+declare function $cycleTimes<T>(n: number, iterable: $SourceIterable<T>): $ResultIterable<T>;
+
+declare function $cycleTimes<T>(n: number): (iterable: $SourceIterable<T>) => $ResultIterable<T>;
+
+export default $cycleTimes;