+++ src/methods/$cycle/$cycle.d.ts
@@ -1,0 +1,5 @@
+import { $SourceIterable, $ResultIterable } from '../../types/$iterable';
+
+declare function $cycle<T>(iterable: $SourceIterable<T>): $ResultIterable<T>;
+
+export default $cycle;