+++ src/methods/$cycle/$cycle.js
@@ -1,0 +1,7 @@
+import $cycleTimes from '../$cycle-times/$cycle-times';
+
+export function $cycle(iterable) {
+  return $cycleTimes(Infinity, iterable);
+}
+
+export default $cycle;