+++ src/methods/$window/window.mjs
@@ -9,14 +9,14 @@
 import { iterableCurry } from '../../internal/iterable';
 import { CircularBuffer, ReadOnlyCircularBuffer } from '../../internal/circular-buffer';
 import { concat } from '../$concat/concat';
-import { repeat } from '../repeat/repeat';
+import { repeatTimes } from '../repeat-times/repeat-times';
 export function* window(source, size, { filler } = {}) {
   const buffer = new CircularBuffer(size);
   const bufferReadProxy = new ReadOnlyCircularBuffer(buffer);
   buffer.fill(filler);
   let index = 0;
 
-  for (const item of concat(source, repeat(size - 1, filler))) {
+  for (const item of concat(source, repeatTimes(size - 1, filler))) {
     buffer.push(item);
 
     if (index + 1 >= size) {