+++ src/methods/repeat-times/repeat-times.d.ts
@@ -1,0 +1,5 @@
+import { ResultIterable } from '../../types/iterable';
+
+declare function repeatTimes<T>(n: number, value: T): ResultIterable<T>;
+
+export default repeatTimes;