+++ src/methods/repeat/repeat.d.ts
@@ -1,7 +1,5 @@
 import { ResultIterable } from '../../types/iterable';
 
-declare function repeat<T>(n: number, item: T): ResultIterable<T>;
-
-declare function repeat<T>(item: T): ResultIterable<T>;
+declare function repeat<T>(value: T): ResultIterable<T>;
 
 export default repeat;